requirements.txt
PyPDF2==3.0.1
python-dotenv==1.1.0
streamlit==1.45.1
google-generativeai==0.3.2

